x64 plugin signing fix

+++ src/host/extensions/signing.cpp
@@ -216,7 +216,7 @@ intercept::cert::signing::security_class intercept::cert::signing::verifyCert(st
 
     CERT_CHAIN_ENGINE_CONFIG ChainConfig;
     //- sizeof(DWORD) removes WIN8 only flag(dwExclusiveFlags) and makes this win7 compatible
-    ChainConfig.cbSize = sizeof(CERT_CHAIN_ENGINE_CONFIG) - sizeof(DWORD);
+    ChainConfig.cbSize = sizeof(CERT_CHAIN_ENGINE_CONFIG) - sizeof(uintptr_t);
     ChainConfig.hRestrictedRoot = nullptr;
     ChainConfig.hRestrictedTrust = nullptr;
     ChainConfig.hRestrictedOther = nullptr;
